requirements.txt
fastapi
httpx
uvicorn
passlib[bcrypt]
databases
asyncpg
sqlalchemy
python-dotenv
PyJWT
psycopg2-binary
pydantic[email]
